refactor: update imports to use shadcn directory in material consumptions pages
